% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{value}

\item{fun}{mean}

\item{alpha}{confidence interval}

\item{cx}{cx values}

\item{...}{}
}
\description{
Title
}
\examples{
rnorm(25, mean = 25, sd = 10)
}
