% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Title}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{size}

\item{iter}{number of iterations}

\item{p}{probability}

\item{...}{}
}
\value{
A graph
}
\description{
Title
}
