% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymult.R
\name{mymult}
\alias{mymult}
\title{Title}
\usage{
mymult(iter = 100, n = 10, p = c(1, 1, 1, 1)/4)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{size}

\item{p}{probablity}
}
\value{
a numeric vector
}
\description{
Title
}
\examples{
mymult(iter=1000,n=10,p=c(1,2,3,4,2)/12)
}
