% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}
}
\value{
display of curve with probability area
}
\description{
Title
}
