% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{iterations}

\item{gamma}{gamma distruvution}

\item{p}{probability}
}
\value{
nd, ns, and discrete and normal graphs
}
\description{
Title
}
